platform/data/ingestion-orch/dags/caiso_ingestion_dag.py
"""
CAISO Settled Price Ingestion DAG
Scheduled to run hourly for settled prices.
"""
from datetime import datetime, timedelta
from airflow import DAG
from airflow.operators.python import PythonOperator
from airflow.utils.dates import days_ago

default_args = {
    'owner': 'market-intelligence',
    'depends_on_past': False,
    'email_on_failure': True,
    'email': ['ops@254carbon.ai'],
    'retries': 3,
    'retry_delay': timedelta(minutes=5),
}


def run_caiso_ingestion(**context):
    """Execute CAISO settled price ingestion."""
    # TODO: Implement CAISO connector
    # Similar structure to MISO connector
    
    import logging
    logger = logging.getLogger(__name__)
    logger.info("CAISO ingestion stub - implement connector")
    
    return 0


with DAG(
    'caiso_settled_ingestion',
    default_args=default_args,
    description='CAISO settled price data ingestion',
    schedule_interval='0 * * * *',  # Every hour
    start_date=days_ago(1),
    catchup=False,
    tags=['ingestion', 'caiso', 'settled'],
) as dag:
    
    ingest_task = PythonOperator(
        task_id='ingest_caiso_settled',
        python_callable=run_caiso_ingestion,
        provide_context=True,
    )

