platform/data/ingestion-orch/requirements.txt
apache-airflow==2.8.0
apache-airflow-providers-postgres==5.10.0
apache-airflow-providers-http==4.7.0
psycopg2-binary==2.9.9
requests==2.31.0

